+++ apis/config/v1beta1/imageconfig_types.go
@@ -17,12 +17,11 @@ limitations under the License.
 package v1beta1
 
 import (
+	"encoding/json"
+
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
-
-// EDIT THIS FILE!  THIS IS SCAFFOLDING FOR YOU TO OWN!
-// NOTE: json tags are required.  Any new fields you add must have json tags for the fields to be serialized.
 
 //+kubebuilder:object:root=true
 //+kubebuilder:subresource:images
@@ -35,6 +34,10 @@ type ImageConfig struct {
 
 	DefaultImages *DefaultImages `json:"defaults,omitempty"`
 
+	ImagePolicy
+}
+
+type ImagePolicy struct {
 	ImageRegistry string `json:"imageRegistry,omitempty"`
 
 	ImagePullSecret corev1.LocalObjectReference `json:"imagePullSecret,omitempty"`
@@ -52,26 +55,86 @@ type Images struct {
 
 	DSEVersions map[string]string `json:"dse,omitempty"`
 
-	SystemLogger string `json:"system-logger"`
-
-	ConfigBuilder string `json:"config-builder"`
+	SystemLogger string `json:"system-logger,omitempty"`
 
 	Client string `json:"k8ssandra-client,omitempty"`
+
+	ConfigBuilder string `json:"config-builder,omitempty"`
+
+	Others map[string]string `json:",inline,omitempty"`
 }
 
+type _Images Images
+
+func (i *Images) UnmarshalJSON(b []byte) error {
+	var imagesTemp _Images
+	if err := json.Unmarshal(b, &imagesTemp); err != nil {
+		return err
+	}
+	*i = Images(imagesTemp)
+
+	var otherFields map[string]interface{}
+	if err := json.Unmarshal(b, &otherFields); err != nil {
+		return err
+	}
+
+	delete(otherFields, CassandraImageComponent)
+	delete(otherFields, DSEImageComponent)
+	delete(otherFields, SystemLoggerImageComponent)
+	delete(otherFields, ConfigBuilderImageComponent)
+	delete(otherFields, ClientImageComponent)
+
+	others := make(map[string]string, len(otherFields))
+	for k, v := range otherFields {
+		others[k] = v.(string)
+	}
+
+	i.Others = others
+	return nil
+}
+
+const (
+	CassandraImageComponent     string = "cassandra"
+	DSEImageComponent           string = "dse"
+	HCDImageComponent           string = "hcd"
+	SystemLoggerImageComponent  string = "system-logger"
+	ConfigBuilderImageComponent string = "config-builder"
+	ClientImageComponent        string = "k8ssandra-client"
+)
+
+type ImageComponents map[string]ImageComponent
+
 type DefaultImages struct {
-	metav1.TypeMeta `json:",inline"`
+	ImageComponents
+}
 
-	CassandraImageComponent ImageComponent `json:"cassandra,omitempty"`
+func (d *DefaultImages) MarshalJSON() ([]byte, error) {
+	// This shouldn't be required, just like it's not with ImagePolicy, but this is Go..
+	return json.Marshal(d.ImageComponents)
+}
 
-	DSEImageComponent ImageComponent `json:"dse,omitempty"`
+func (d *DefaultImages) UnmarshalJSON(b []byte) error {
+	d.ImageComponents = make(map[string]ImageComponent)
+	var input map[string]json.RawMessage
+	if err := json.Unmarshal(b, &input); err != nil {
+		return err
+	}
 
-	HCDImageComponent ImageComponent `json:"hcd,omitempty"`
+	for k, v := range input {
+		var component ImageComponent
+		if err := json.Unmarshal(v, &component); err != nil {
+			return err
+		}
+		d.ImageComponents[k] = component
+	}
+
+	return nil
 }
 
 type ImageComponent struct {
 	Repository string `json:"repository,omitempty"`
 	Suffix     string `json:"suffix,omitempty"`
+	ImagePolicy
 }
 
 func init() {